+++ app/controllers/ru_captcha/captcha_controller.rb
@@ -3,11 +3,13 @@ module RuCaptcha
     def index
       headers['Cache-Control'] = 'no-cache, no-store, max-age=0, must-revalidate'
       headers['Pragma'] = 'no-cache'
+      data = generate_rucaptcha
+      opts = { disposition: 'inline', type: 'image/gif' }
 
       if Gem.win_platform?
-        send_file generate_rucaptcha, disposition: 'inline', type: 'image/png'
+        send_file data, opts
       else
-        send_data generate_rucaptcha, disposition: 'inline', type: 'image/png'
+        send_data data, opts
       end
     end
   end