+++ lib/rucaptcha.rb
@@ -1,12 +1,12 @@
 require 'rails'
 require 'action_controller'
 require 'active_support/all'
+require_relative 'rucaptcha/rucaptcha'
 require_relative 'rucaptcha/version'
 require_relative 'rucaptcha/configuration'
 require_relative 'rucaptcha/controller_helpers'
 require_relative 'rucaptcha/view_helpers'
 require_relative 'rucaptcha/cache'
-require_relative 'rucaptcha/captcha'
 require_relative 'rucaptcha/engine'
 
 module RuCaptcha
@@ -14,12 +14,8 @@ module RuCaptcha
     def config
       return @config if defined?(@config)
       @config = Configuration.new
-      @config.len         = 4
-      @config.font_size   = 45
-      @config.implode     = 0.3
-      @config.cache_limit = 100
-      @config.expires_in  = 2.minutes
       @config.style       = :colorful
+      @config.expires_in  = 2.minutes
       if Rails.application
         @config.cache_store = Rails.application.config.cache_store
       else
@@ -30,6 +26,11 @@ module RuCaptcha
 
     def configure(&block)
       config.instance_exec(&block)
+    end
+
+    def generate()
+      style = config.style == :colorful ? 1 : 0
+      self.create(style)
     end
   end
 end